Create Frequency dictionary
When Anton finished reading "War and Peace", he decided to find out the number of specific words used in the book.

Help Anton write a simplified version of such a program, that will be capable of counting the words separated with space, and print the statistics.

The program must prompt a user for 1 line and print out each unique word with the number of its usages in the line in the following format "word amount" (case insensitive!). The word order does not matter, each word must be printed only once.