Create Python code: Text formatting:Markdown style
Implement a separate function for each of the formatters. It will keep your code structured. With functions, you also will be able to find and fix a bug with ease if something is wrong.

The program should work in the following way:

Ask a user to input a formatter.
If the formatter doesn't exist, print the following error message: Unknown formatting type or command
Ask a user to input a text that will be applied to the formatter: Text: <user's input>.
Save the text with the chosen formatter applied to it and print the markdown. Each time you should print out the whole text in markdown accumulated so far.
Different formatters may require different inputs.

The new-line formatter does not require text input.

Plain, bold, italic, and inline-code formatters require only text input, and should not add an extra space or line break at the end:

	
Text: > Some text

Headers require a level and text. A level is a number from 1 to 6. Don't forget to check it too: if it is out of bounds, print the corresponding error: The level should be within the range of 1 to 6. Also, remember that a heading automatically adds a new line in the end.

	
Level: > 4
Text: > Hello World!

Link requires a label and a URL:

	
Label: > Tutorial
URL: > https://www.markdownguide.org/basic-syntax/