Create 5:Create dictionaries
Work on project. Stage 1/5: Create dictionaries
Project: Food Blog Backend
 Hard  41 minutes 
175 users solved this problem. Latest completion was 
2 minutes ago
.
Description
Your great-grandmother asked you to copy "to these computers" all the recipes that she had been collecting in her notebook for several decades. You think that this task can be a good opportunity to practice your skills and build a tool that would collect the data in a database. It will allow you to create a database of recipes that you will be able to use in the blog with the great-grandma's recipes. You started by writing down the names of the ingredients and measures that Grandma used. It also may be good to assign different dishes to different times of the day. It's time to create a database and dictionaries!

Below is a diagram of the database tables. PK means the Primary Key.



Theory
SQLite3 is a relational database management system. You can use it for free; it is licensed as Public Domain. The system implements most of the SQL standards. For this project, we will use the basic SQLite3 functions.

First, import the sqlite3 library:

import sqlite3
To connect your script to an SQLite database, use the connect() method from the sqlite3 library and create a database cursor with the cursor() method:

conn = sqlite3.connect(data_base_name)
cursor_name = conn.cursor()
To execute an SQL query, use the execute() cursor method. This method will return an object that contains the result of the query. For example:

result = cursor_name.execute(SQL_query_as_string)
Another two important methods are close() and commit(). Remember that you need to confirm the SQL queries with the commit command. Otherwise, the data will be lost. At the end of your code, disconnect your database. Both methods are related to the database connection:

conn.commit()
conn.close()
If you need more information, the SQLite Tutorial will help you!

Objectives
Create a database. Pass the name of the database to the script as an argument.
Create a table named meals with two columns: meal_id of an integer type with the primary key attribute, and meal_name of a text type and with the unique and not null attribute.
Create a table named ingredients with two columns: ingredient_id of an integer type with the primary key attribute and ingredient_name of a text type with the unique and not null attribute. Multi-word ingredients are out of scope, you don't need to implement their support in your script.
Create a table named measures with two columns: measure_id of an integer type with the primary key attribute, and measure_name of a text type with the unique attribute.
Populate the tables. Those tables are the dictionaries for the Food Blog system, you need to fill them once for the rest of the stages.
data = {"meals": ("breakfast", "brunch", "lunch", "supper"),
        "ingredients": ("milk", "cacao", "strawberry", "blueberry", "blackberry", "sugar"),
        "measures": ("ml", "g", "l", "cup", "tbsp", "tsp", "dsp", "")}
Tests do not check the output. You can print anything you want. Tests will check only the database file that your script will create.
Do not add other items to the dictionaries. This may affect the test results in this and the next stages