Create 5: Time for recipes!
Create a table named recipes with three columns: recipe_id of an integer type with the primary key attribute, recipe_name of a text type with the not-null attribute, and recipe_description of a text type.
Prepare a simple system that allows you to populate this table. Ask for the recipe name and the cooking directions, and insert the data into the table.
When a zero-length string is entered for the recipe name the script should terminate. Remember to commit your changes and close the database.
Remember! You can print anything you want. Tests will check only the database file that your script will create and populate.
You do not need to validate the entered data. The tests will pass the correct values.