Create 5: Too many ingredients!
Create a table named quantity with five columns: quantity_id of an INTEGER type with the PRIMARY KEY attribute, and four other columns: measure_id, ingredient_id, quantity and recipe_id . They should be of an INTEGER type with the NOT NULL attribute.
Assign the following columns measure_id, ingredient_id and recipe_id as Foreign Keys to the following tables (columns): measures (measure_id), ingredients (ingredient_id), and recipes (recipe_id)
After asking a user about certain mealtime, make a loop that will gather information about the ingredients. The ingredients should be entered in the following format: quantity measure ingredient.
Pressing <Enter> should finish the information gathering.
The measure parameter should start with a string provided by a user. If there is more than one measure that starts with the provided string, ask the user again. For example tbs and tbsp both start with the t. So the 1 t sugar entry should not pass.
Mind that the measures table contains an entry where the measure_name is empty string, it means, that the measure could be not provided. In this case, use this entry to relate tables. For example, 1 strawberry should have a measure_key from the entry with an empty name.
The ingredient parameter should contain strings provided by a user. If there is more than one ingredient that contains the provided string, ask the user again. For example strawberry and blueberry both contain berry as part of the string. So the 10 kg berry entry should not pass.
Tests do not check the output. You can print anything you want. Tests will check only the database file that your script will create.